undefined-checking skipframes
the way it it works currently it's not really optional. i noticed a behavior change upgrading from 2.6.2, where i would leave off the last 3 args to `cache.addSpriteSheet` and it worked fine -- now in 2.7.2 need to zero them all out as that func calls this one leaving `skipFrames` undefined, so when it's added to `frameMax` it's NaN and everything goes to hell.