setText updates
BitmapText.setText will check if the value given is falsey but not a zero and set to an empty string if so.
BitmapText.setText will now cast the given value to a string before setting.
BitmapText.setText will not change the text via `setText` unless the new text is different to the old one.