Pass game instance as boot callback arguments.
It may be useful to run some operations while the game is booting.
However, if a living instance of the game is necessary during this
process, only way possible is using a global variable at the moment.

This commit aims to amend that so whenever needed the game instance
reference is passed down and can be used inside the {pre,post}Boot
callbacks.