2d Camera startFollow should respect useBounds
It was possible for the 2d camera's `startFollow` method to set the `scrollX` and `scrollY` values out of the camera's bounds, even when `useBounds` was set to true.

This PR ensures that `useBounds` is respected by the `startFollow` method when setting the `scrollX` and `scrollY` values. 

Question - should this check also happen when setting the `midpoint` with the `fx` and `fy` values?